moved the config.go file in internal to the pkg folder
